Sync back to the previous tenkft-stable version
After attempting to find the commit that syncs with this version, and failing to get that quickly, this commit copies the files from the previous installed gem on our production servers into our stable branch. This creates a point in time snapshot of what we had in the previous, deleted tenkft-stable branch.

Note that picking up newer changes from onelogin/ruby-saml will have to go into a new branch as this commt pollutes commit history and we can no longer cleanly merge into this branch