Namespace XMLSecurity --> XMLSecurityFrozen081
Rename folder path and update module and file names as needed to get a rails console to load and be able to access the original and frozen versions of the modules and classes. The require statements will get tripped when there are two gems with the same paths, so we need unique paths for require to isolate the two gem sources.